refactor StackModal - extract StackFrequencyAndDates - Rename StackOrdersProgress and move related orders componentd there
